feat: remove Command(goto=Node) + rename missing_mdetails to prompt_missing_details
