+++ src/meetingmuse/models/state.py
@@ -35,5 +35,8 @@ class MeetingMuseBotState(BaseModel):
     # Information about the meeting being scheduled
     meeting_details: MeetingFindings = Field(default_factory=MeetingFindings)
 
-    # Human input
+    # Prompt input for Human nodes
     ai_prompt_input: Optional[str] = None
+
+    # Human input
+    human_input: Optional[str] = None
